// src/app/Core/Models/addUser.model.ts
export interface AddUserDto {
  id?: number;
  fNameAr: string;
  secNameAr: string;
  thirdNameAr: string;
  lNameAr: string;
  fNameEn: string;
  secNameEn: string;
  thirdNameEn: string;
  lNameEn: string;
  email: string;
  countryCode: string;
  nationalId: number;
  mobileNumber: string;
  birthDate: string;
  gender: number;
  maritalStatus: number;
  addressAr: string;
  addressEn: string;
  jobId: number;
  departmentId: number;
}